+++ tienda/src/main/java/com/mintic/tienda/entity/TipoPrenda.java
@@ -15,16 +15,16 @@ public class TipoPrenda {
     @Id
     @GeneratedValue( strategy=GenerationType.AUTO )
     @Column(name = "id_tipo_prenda")
-    private int idTipoPrenda;
+    private Long idTipoPrenda;
     
     @Column(name = "tipo")
     private String tipo;
 
-    public int getIdTipoPrenda() {
+    public Long getIdTipoPrenda() {
         return idTipoPrenda;
     }
 
-    public void setIdTipoPrenda(int idTipoPrenda) {
+    public void setIdTipoPrenda(Long idTipoPrenda) {
         this.idTipoPrenda = idTipoPrenda;
     }
