+++ tienda/src/main/java/com/mintic/tienda/repository/ITallaRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.mintic.tienda.entity.Talla;
 
 @Repository
-public interface ITallaRepo extends JpaRepository<Talla, Integer>{
+public interface ITallaRepo extends JpaRepository<Talla, Long>{
     
 }
