+++ tienda/src/main/java/com/mintic/tienda/service/DTO/MarcaDTO.java
@@ -3,18 +3,21 @@ package com.mintic.tienda.service.DTO;
 
 public class MarcaDTO {
     
-    private int idMarca;
+    private Long idMarca;
     private String descripcion;
     
-    public int getIdMarca() {
+    public Long getIdMarca() {
         return idMarca;
     }
-    public void setIdMarca(int idMarca) {
+
+    public void setIdMarca(Long idMarca) {
         this.idMarca = idMarca;
     }
+
     public String getDescripcion() {
         return descripcion;
     }
+    
     public void setDescripcion(String descripcion) {
         this.descripcion = descripcion;
     }
